+++ store.js
@@ -14,8 +14,12 @@ import * as dashboardDetail from 'components/dashboards/detail/dashboard-detail'
 import * as dashboardThumbnailList from 'components/dashboards/thumbnail-list/dashboard-thumbnail-list';
 import * as widgetBlockModule from 'components/dashboards/wysiwyg/widget-block/widget-block';
 import * as widgetBlockEditionModule from 'components/dashboards/wysiwyg/widget-block-edition/widget-block-edition';
-import * as exploreDatasetFilters from 'components/app/explore/explore-dataset-filters/explore-dataset-filters';
+
+// Dataset
 import * as similarDatasets from 'components/app/explore/similar-datasets/similar-datasets';
+
+// Explore
+import * as exploreDatasetFilters from 'components/app/explore/explore-dataset-filters/explore-dataset-filters';
 
 // Widget editor
 import { reducers as widgetEditorModules } from 'widget-editor';
@@ -44,6 +48,8 @@ const reducer = combineReducers({
 
   // Explore
   exploreDatasetFilters: handleModule(exploreDatasetFilters),
+
+  // Explore detail
   similarDatasets: handleModule(similarDatasets)
 });
